iot-simulator/requirements.txt
python-dotenv==1.0.0
paho-mqtt==1.6.1